+++ README.md
@@ -38,16 +38,7 @@ $ pip install 'incydr[all]'.
 ```
 
 For more details see the [SDK Documentation](docs/sdk/index.md) and the [CLI Documentation](docs/cli/getting_started.md).
-.
+
 ## License
 
 `incydr` is distributed under the terms of the [MIT](https://spdx.org/licenses/MIT.html) license.
-
------
-## Getting Started
-
-```python
-from incydr import Client
-
-client = Client(url="api-gateway-url.code42.com", api_client_id="key-42", api_client_secret="c42")
-```
