+++ pyproject.toml
@@ -44,7 +44,7 @@ Source = "https://github.com/code42/incydr_python/incydr"
 incydr = "incydr.cli:incydr"
 
 [tool.hatch.version]
-path = "src/incydr/__about__.py"
+path = "src/_client/version.py"
 
 [tool.hatch.build.targets.sdist]
 [tool.hatch.build.targets.wheel]
@@ -92,9 +92,8 @@ no-cov = "cov --no-cov"
 [tool.coverage.run]
 branch = true
 parallel = true
-omit = [
-  "incydr/__about__.py",
-]
+omit = []
+
 [tool.coverage.report]
 exclude_lines = [
   "no cov",
