--- src/_client/__about__.py
@@ -1,4 +1,0 @@
-# SPDX-FileCopyrightText: 2022-present Code42 Software <integrations@code42.com>
-#
-# SPDX-License-Identifier: MIT
-__version__ = "0.1.0"
