+++ src/incydr/__init__.py
@@ -3,17 +3,12 @@
 # SPDX-License-Identifier: MIT
 from _client import __version__
 from _client.core.client import Client
-from _client.models import *  # noqa
 from _client.queries.alerts import AlertQuery
 from _client.queries.file_events import EventQuery
 from _client.utils import *  # noqa
+from incydr.models import *  # noqa
 
-__all__ = [
-    "__version__",
-    "Client",
-    "AlertQuery",
-    "EventQuery"
-]
+__all__ = ["__version__", "Client", "AlertQuery", "EventQuery"]
 
 __locals = locals()
 for __name in __all__:
