+++ src/incydr/enums/__init__.py
@@ -1,7 +1,3 @@
-from _client.enums import alerts
-from _client.enums import cases
-from _client.enums import devices
-from _client.enums import file_events
 from _client.enums import SortDirection
-from _client.enums import trusted_activities
-from _client.enums import watchlists
+
+__all__ = ["SortDirection"]
