+++ src/incydr/enums/file_events.py
@@ -1,0 +1,21 @@
+from _client.enums.file_events import Category
+from _client.enums.file_events import EventAction
+from _client.enums.file_events import EventSearchTerm
+from _client.enums.file_events import FileCategory
+from _client.enums.file_events import ReportType
+from _client.enums.file_events import RiskIndicators
+from _client.enums.file_events import RiskSeverity
+from _client.enums.file_events import ShareType
+from _client.enums.file_events import TrustReason
+
+__all__ = [
+    "EventSearchTerm",
+    "FileCategory",
+    "EventAction",
+    "Category",
+    "ShareType",
+    "ReportType",
+    "RiskIndicators",
+    "TrustReason",
+    "RiskSeverity",
+]
