+++ src/incydr/enums/trusted_activities.py
@@ -1,0 +1,15 @@
+from _client.enums.trusted_activities import ActivityType
+from _client.enums.trusted_activities import CloudShareApps
+from _client.enums.trusted_activities import CloudSyncApps
+from _client.enums.trusted_activities import EmailServices
+from _client.enums.trusted_activities import PrincipalType
+from _client.enums.trusted_activities import SortKeys
+
+__all__ = [
+    "ActivityType",
+    "CloudSyncApps",
+    "CloudShareApps",
+    "EmailServices",
+    "PrincipalType",
+    "SortKeys",
+]
