+++ src/incydr/enums/watchlists.py
@@ -1,0 +1,3 @@
+from _client.enums.watchlists import WatchlistType
+
+__all__ = ["WatchlistType"]
