+++ tests/queries/test_event_query.py
@@ -4,9 +4,8 @@ from datetime import timedelta
 import pytest
 from _client.queries.file_events import Filter
 from _client.queries.file_events import FilterGroup
-from pydantic import ValidationError
-
 from incydr import EventQuery
+from pydantic import ValidationError
 
 TEST_START_DATE = "P1D"
 TEST_TIMESTAMP = "2020-09-10 11:12:13"
