+++ tests/test_directory_groups.py
@@ -4,9 +4,8 @@ from urllib.parse import urlencode
 from _client.core.client import Client
 from _client.directory_groups.models import DirectoryGroup
 from _client.directory_groups.models import DirectoryGroupsPage
-from pytest_httpserver import HTTPServer
-
 from incydr.cli import incydr
+from pytest_httpserver import HTTPServer
 
 
 def test_get_page_when_default_params_returns_expected_data(
