+++ tests/test_file_events.py
@@ -14,9 +14,8 @@ from _client.file_events.models.response import SavedSearch
 from _client.file_events.models.response import SearchFilter
 from _client.file_events.models.response import SearchFilterGroup
 from _client.queries.file_events import EventQuery
-from pytest_httpserver import HTTPServer
-
 from incydr.cli import incydr
+from pytest_httpserver import HTTPServer
 
 MICROSECOND_FORMAT = "%Y-%m-%dT%H:%M:%S.%fZ"
 TEST_EVENT_1 = {
